Addressed reviewer comments
